admini update

+++ controllers/logoController.js
@@ -1,5 +1,6 @@
 const fs = require("fs");
 const Logo = require("../models/logoModel");
+const SellerLogo = require("../models/admin/sellerLogo.model");
 
 exports.addLogo = async (req, res) => {
   try {
@@ -12,6 +13,38 @@ exports.addLogo = async (req, res) => {
     }
 
     const result = await Logo.create({ logo: logo });
+
+    if (!result) {
+      return res.status(404).json({
+        success: false,
+        error: "Logo not added",
+      });
+    }
+
+    res.status(200).json({
+      success: true,
+      message: "Logo added successfully",
+      data: result,
+    });
+  } catch (error) {
+    res.status(500).json({
+      success: false,
+      error: error.message,
+    });
+  }
+};
+
+exports.addSellerLogo = async (req, res) => {
+  try {
+    const logo = req?.file?.filename;
+    if (!logo) {
+      return res.status(404).json({
+        success: false,
+        error: "Logo is requred",
+      });
+    }
+
+    const result = await SellerLogo.create({ logo: logo });
 
     if (!result) {
       return res.status(404).json({
@@ -57,6 +90,30 @@ exports.getLogos = async (req, res) => {
   }
 };
 
+exports.getSellerLogos = async (req, res) => {
+  try {
+    const logo = await SellerLogo.find({});
+
+    if (!logo) {
+      return res.status(404).json({
+        success: false,
+        error: "Logo not found",
+      });
+    }
+
+    res.status(200).json({
+      success: true,
+      message: "Logo found successfully",
+      data: logo,
+    });
+  } catch (error) {
+    res.status(500).json({
+      success: false,
+      error: error.message,
+    });
+  }
+};
+
 exports.updateLogo = async (req, res) => {
   try {
     const logo = req?.file?.filename;
@@ -72,6 +129,42 @@ exports.updateLogo = async (req, res) => {
 
     if (isLogo) {
       await Logo.findByIdAndUpdate(id, { logo: logo }, { new: true });
+
+      fs.unlink(`./uploads/logo/${isLogo?.logo}`, (err) => {
+        if (err) {
+          console.error(err);
+          return;
+        }
+      });
+
+      res.status(200).json({
+        success: true,
+        message: "Logo updated successfully",
+      });
+    }
+  } catch (error) {
+    res.status(500).json({
+      success: false,
+      error: error.message,
+    });
+  }
+};
+
+exports.updateSellerLogo = async (req, res) => {
+  try {
+    const logo = req?.file?.filename;
+    if (!logo) {
+      return res.status(400).json({
+        success: false,
+        error: "Logo is required",
+      });
+    }
+
+    const id = req?.params?.id;
+    const isLogo = await SellerLogo.findOne({ _id: id });
+
+    if (isLogo) {
+      await SellerLogo.findByIdAndUpdate(id, { logo: logo }, { new: true });
 
       fs.unlink(`./uploads/logo/${isLogo?.logo}`, (err) => {
         if (err) {
